+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- golang: `sync` delete `dep` installed by `go get ...`
+
 - golang: `sync`/`update` grab favourite packages listed in TOML
 
 - golang: `sync`/`update` grab linters with [gometalinter](https://github.com/alecthomas/gometalinter)